requirements.txt
requests>=2.31.0
Pillow>=10.0.0
weasyprint>=62.0
python-dotenv>=1.0.0

# LIFI Agent Toolkit
flask>=3.0
gspread>=6.0
google-auth>=2.0
APScheduler>=3.10
google-api-python-client>=2.0
pywebpush>=2.0.0
gunicorn>=21.2
cryptography>=41.0
